+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Column comments will now be properly added to the Avro schema in struct fields
+nested within arrays
+
 ## [1.5.3] 2021-02-12
 
 ### Added